+++ app/CHANGELOG.md
@@ -2,8 +2,11 @@ CHANGELOG:
 
 ## Version 1.2.0
 
+- Add: Dashboard allows forms sync
+- Add: Dashboard configures different settings
 - Update: Adjust email 
 - Update: Adjust schedulte templates for new Head Teaching Assistant
+- Fix: 'Schedule Template' now is hidden during rotation of schedules
 
 ## Version 1.1.3